+++ R/dirknn.R
@@ -6,10 +6,10 @@ dirknn <- function(xnew, x, y, k, type = "C", parallel = FALSE) {
 
 #[export]
 dirknn.cv <- function(y, x, k = 5:10, type = "C", folds = NULL, nfolds = 10, 
-stratified = TRUE, seed = FALSE, parallel = FALSE, pred.ret = FALSE) {
+stratified = TRUE, seed = NULL, parallel = FALSE, pred.ret = FALSE) {
   
   crit <- matrix(nrow = nfolds, ncol = length(k))
-  if ( is.null( folds ) )   folds <- makefolds(y, nfolds = nfolds, stratified = stratified, seed = FALSE)
+  if ( is.null( folds ) )   folds <- .makefolds(y, nfolds = nfolds, stratified = stratified, seed = NULL)
   preds <- list() 
   y <- as.numeric(y)
   for (i in 1:nfolds) {  
@@ -30,40 +30,3 @@ stratified = TRUE, seed = FALSE, parallel = FALSE, pred.ret = FALSE) {
 }
 
 
-
-makefolds <- function(ina, nfolds = 10, stratified = TRUE, seed = FALSE) {
-  names <- paste("Fold", 1:nfolds)
-  runs <- sapply(names, function(x) NULL)
-  if (seed)  set.seed(1234)
-
-  if ( !stratified ) {
-    oop <- options(warn = -1)
-    on.exit(options(oop))
-    ep <- sample( length(ina) )
-    nr <- round( length(ina)/nfolds )
-    mat <- matrix( ep[1:(nr * nfolds) ], ncol = nfolds )
-    mat[ -c( 1:length(ina) ) ] <- NA
-    for ( i in 1:nfolds ) runs[[ i ]] <- mat[, i]
-    rem <- ep[ - c(1:(nr * nfolds)) ]
-    ela <- sample(nfolds, length(rem))
-    if ( length(ela) > 0 )  for ( i in 1:length(ela) )  runs[[ ela[i] ]] <- c( runs[[ i ]], rem[ i ] )
-  } else {
-    labs <- unique(ina)
-    run <- list()
-    for (i in 1:length(labs)) {
-      names <- which( ina == labs[i] )
-      run[[i]] <- sample(names)
-    }
-    run <- unlist(run)
-    for ( i in 1:length(ina) ) {
-      k <- i %% nfolds
-      if ( k == 0 )  k <- nfolds
-      runs[[k]] <- c( runs[[ k ]], run[i] )
-    }
-  }
-  for (i in 1:nfolds)  {
-    if ( any( is.na(runs[[ i ]]) ) )  runs[[ i ]] <- runs[[ i ]][ !is.na(runs[[ i ]]) ]
-  }
-  if ( length(runs[[ nfolds ]]) == 0 ) runs[[ nfolds ]] <- NULL
-  runs
-}