+++ R/regression_models.R
@@ -58,39 +58,41 @@ gammareg <- function(y, x, tol = 1e-07, maxiters = 100) {
   X <- model.matrix( y~., data.frame(x) )
   sx <- Rfast::colsums(X)
   dm <- dim(X)
-  p <- dm[2]
-  n <- dm[1]
+  n <- dm[1]  ;  p <- dm[2]
+
   mod <- Rfast::gammacon(y, tol = tol) 
-  m <- mod$be
-  be <- c( m, numeric(p - 1) )
-  d1 <-  - 0.5 * mod$deviance - n
-  con <- y * m
-  com <- con * X
-  der <-  - Rfast::colsums(com) + sx
-  der2 <- crossprod(com, X)
+  m <- exp(mod$be)
+  be <- c( mod$be, numeric(p - 1) )
+  d1 <-  - sum(y / m) + n * mod$be
+  con <- y / m
+  der <- Rfast::eachcol.apply(X, con) - sx
+  der2 <-  - crossprod( con * X, X) 
   be <- be - solve(der2, der)
-  m <- as.vector( exp( - X %*% be ) )
-  con <- y * m
-  d2 <- sum( log(con) - con )
+  xbe <-  X %*% be
+  m <- as.vector( exp( xbe ) )
+  con <- y / m
+  d2 <-  - sum(con) - sum(xbe)
+
   i <- 2
-  while ( abs(d2 - d1) > tol  & i < maxiters) {
+  while ( abs(d1 - d2) > tol  & i < maxiters) {
     i <- i + 1
     d1 <- d2
-    com <- con * X
-    der <-  - Rfast::colsums(com) + sx
-    der2 <- crossprod(com, X)
+    der <- Rfast::eachcol.apply(X, con) - sx
+    der2 <-  - crossprod( con * X, X) 
     be <- be - solve(der2, der)
-    m <- as.vector( exp( - X %*% be ) )
-    con <- y * m
-    d2 <- sum( log(con) - con )
+    xbe <-  X %*% be
+    m <- as.vector( exp( xbe ) )
+    con <- y / m
+    d2 <-  - sum(con) - sum(xbe)
   }
   phi <- sum( (con - 1)^2 ) / (n - p)
-  devi <-  - 2 * d2 - 2 * n
+  devi <-  - 2 * sum( log(con) ) + sum(con) - n
   info <- c(i, devi, phi)
   names(info) <- c("iters", "deviance", "phi")
   names(be) <- colnames(X)
   list(info = info, be = be)
 }
+
 
 
 #[export]
@@ -254,7 +256,7 @@ multinom.reg <- function(y, x, tol = 1e-07, maxiters = 50) {
       }  
       b2 <- b1 + solve(der2, der)
     }, silent = TRUE)
-    if (class(res) == "try-error")   b2 <- b1
+    if (inherits(res,"try-error"))   b2 <- b1
     colnames(b2) <- paste("Y", 1:d, sep = "")
     rownames(b2) <- colnames(X)
     Y1 <- design_matrix(y, ones = FALSE)
@@ -383,6 +385,8 @@ spml.reg <- function(y, x, tol = 1e-07, seb = FALSE, maxiters = 100) {
   x <- model.matrix(~., data.frame(x) )
   y <- as.matrix(y)
   mod <- .Call(Rfast_spml_reg, y, x, tol, seb, maxiters)
+  colnames(mod$be) <- c("Cosinus of y", "Sinus of y")
+  rownames(mod$be) <- colnames(x)   
  
   if ( seb ) {
     colnames(mod$seb) <- c("Cosinus of y", "Sinus of y")