+++ src/maha_ex.cpp
@@ -1,4 +1,4 @@
-﻿// [[Rcpp::depends(RcppArmadillo)]]
+// [[Rcpp::depends(RcppArmadillo)]]
 #include <RcppArmadillo.h>
 using namespace Rcpp;